man: ip-link.8: add a note for gso_ipv4_max_size
As Paolo noticed, a skb->len check against gso_max_size was added in:

  https://lore.kernel.org/netdev/20231219125331.4127498-1-edumazet@google.com/

gso_max_size needs to be set to a value greater than or equal to
gso_ipv4_max_size to make BIG TCP IPv4 work properly.

To not break the current setup, this patch just adds a note into its
man doc for this.

Reported-by: Xiumei Mu <xmu@redhat.com>
Signed-off-by: Xin Long <lucien.xin@gmail.com>